Fixed issue where command compiler was not able to use System.IO and other namespaces Added script tester to diagnose C# script commands Broke out editor state reporting and command execution into their own files Removed batch command and udon script support for now Created sequence diagram to illustrate the communication flow between user, MCP Client, MCP Server, and Unity Editor.
